SB-331: Add a maven invoker test case for missing artifacts
Implemented.

Note: This can't be done at a POM-level because the settings.xml file for the Maven invoker tests is pointing to the local file system and, for the sake of quicker testing, it's not actually hitting a remote repository for resolving artifacts, (it's only hitting it when deploying them).